[fix] : 서비스 간 순환참조 방지를 위해 메소드 이동
